feat(core): add modules & components folders
