+++ src/components/Button.vue
@@ -1,0 +1,11 @@
+<template>
+  <button class="bg-green-500 px-5 py-2 text-white rounded">
+    <slot />
+  </button>
+</template>
+
+<script>
+export default {
+  name: 'Button',
+};
+</script>
