+++ src/main.js
@@ -1,6 +1,11 @@
 import { createApp } from 'vue';
 
+import Button from '@/components/Button.vue';
+
 import App from './App.vue';
 import router from './router';
 
-createApp(App).use(router).mount('#app');
+const app = createApp(App);
+app.component('Button', Button);
+app.use(router);
+app.mount('#app');
