+++ src/pages/GitView.vue
@@ -1,67 +1,24 @@
 <template>
-  <div class="flex h-full flex-col">
-    <div class="flex justify-between py-4 px-4">
-      <router-link to="/" class="bg-green-500 px-5 py-2 text-white rounded"
-        >Back</router-link
-      >
-      <div>{{ folderPath }}</div>
-      <div></div>
-    </div>
-    <div class="w-full flex h-full">
-      <div class="flex self-center line relative">
-        <div
-          v-for="{ sha, author, isHead } in commits"
-          :key="sha"
-          class="px-4 text-center"
-        >
-          <div
-            class="bg-gray-300 rounded-1/2 w-24 h-24"
-            :class="{ 'border-4 border-green-500': isHead }"
-          ></div>
-          <p class="pt-2">{{ sha }}</p>
-          <p class="pt-2">{{ author }}</p>
-          <p class="pt-2" v-if="isHead">HEAD</p>
-        </div>
-      </div>
-    </div>
-  </div>
+  <GitViewModule :folder-path="folderPath" />
 </template>
 
 <script>
-import { ipcRenderer } from 'electron';
-import { onBeforeMount, ref } from 'vue';
 import { useRoute } from 'vue-router';
 
+import GitViewModule from '@/modules/GitView';
+
 export default {
+  name: 'GitViewPage',
+
+  components: {
+    GitViewModule,
+  },
+
   setup() {
     const route = useRoute();
     const { folderPath } = route.query;
-    const commits = ref([]);
 
-    const resolveGetGitLogs = () => {
-      ipcRenderer.on('getGitLogs-reply', (event, args) => {
-        commits.value = args;
-      });
-    };
-    resolveGetGitLogs();
-    onBeforeMount(() => {
-      ipcRenderer.send('getGitLogs-event', folderPath);
-    });
-
-    return { commits, folderPath };
+    return { folderPath };
   },
 };
 </script>
-
-<style>
-.line::before {
-  content: '';
-  height: 2px;
-  width: 100%;
-  background-color: black;
-  position: absolute;
-  top: 40%;
-  z-index: -1;
-  transform: translate(-1rem, 0px);
-}
-</style>
